+++ src/app/components/new/login-btn.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { signIn } from "next-auth/react";
+
+import { Button } from "@/components/ui/button";
+
+export default function LoginBtn() {
+  return (
+    <Button className="w-full" onClick={() => signIn()}>
+      ログイン
+    </Button>
+  );
+}
